Move a couple of helpers to the remove project
